requirements.txt
git+https://github.com/Rapptz/discord.py
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.3.0
cffi==1.13.1
chardet==3.0.4
discord.py==1.3.1
idna==2.8
multidict==4.5.2
pycparser==2.19
PyNaCl==1.3.0
six==1.12.0
virtualenv==16.7.7
websockets==6.0
yarl==1.3.0
opuslib==3.0.1
pillow==7.0.0